Add an explicit cast to shut up g++47 on FreeBSD

+++ neo/idlib/sys/posix/posix_thread.cpp
@@ -103,7 +103,14 @@ Sys_GetCurrentThreadID
 */
 uintptr_t Sys_GetCurrentThreadID()
 {
-	return pthread_self();
+	/*
+	 * This cast is save because pthread_self()
+	 * returns a pointer and uintptr_t is
+	 * designed to hold a pointer. The compiler
+	 * is just too stupid to know. :)
+	 *  -- Yamagi
+	 */
+	return ( uintptr_t )pthread_self();
 }
 
 /*
